Split downloader and converter for binary evolution models
